Download and import OSM data on new project creation
A new popup suggests to download and import OSM data if the opened project
has no features in it.

The user can select a subset of OSM features to be downloaded, or by
default it downloads everything related.

Additional option is to automatically download everything into the
ALAQS database.

All OSM mapping configuration is in `alaqs_config:LAYERS_CONFIG`.