Fix movement calculation
The profile and engine will calculate the pollutant emissions correctly, but the Emissions geometry will be incorrect.
This is because the Profile shows the path of the airplane ignoring the azimuth of the Runway
and it's geometry is stored precalculated with the Runway in the resulting FlightEmissions object.
However, the geometry needs to be rotated to match the respective Runway of each Movement.